+++ FrameWindow.cpp
@@ -1,22 +1,5 @@
 #include <GLFW/glfw3.h>
-#include "fwMainFrame.h"
-
 #include "FrameWindow.h"
-
-fwFrame* fwInitFrameWindow(GLFWwindow* window, int rowCount, int columnCount,
-	const std::initializer_list<int>& widths,
-	const std::initializer_list<int>& heights)
-{
-	void fwInitCallbacks(GLFWwindow * window);
-	int x, y, w, h;
-	glfwGetWindowPos(window, &x, &y);
-	glfwGetWindowSize(window, &w, &h);
-	fwMainFrame* mf = new fwMainFrame(x, y, w, h, window, rowCount, columnCount, widths, heights);
-	glfwSetWindowUserPointer(window, mf);
-
-	fwInitCallbacks(window);
-	return mf;
-}
 
 fwFrame* fwInitWidgetAsFrame(fwFrame* parent, int row, int column, int rowCount, int columnCount, const std::initializer_list<int>& widths, const std::initializer_list<int>& heights)
 {
@@ -27,17 +10,17 @@ fwFrame* fwInitWidgetAsFrame(fwFrame* parent, int row, int column, int rowCount,
 	return newFrame;
 }
 
-bool fwInitWidgetAsLeaf(fwFrame* parent, int row, int column, fwLeafWidget* leaf)
+fwWidget* fwSetWidgetAsLeaf(fwFrame* parent, int row, int column, fwLeafWidget* leaf)
 {
-	if (parent->getElement(row, column) != nullptr) return false;
+	fwWidget* oldElement = parent->getElement(row, column);
 	parent->replaceElement(row, column, leaf);
-	return true;
+	return oldElement;
 }
 
-void fwUpdate(fwFrame* mainFrame)
+void fwUpdate(fwMainFrame* mainFrame)
 {
 	if(mainFrame->type() == fwWidget::WidgetType::MainFrame)
-		((fwMainFrame*)mainFrame)->display();
+		((fwMainFrame*)mainFrame)->onRepaint();
 }
 
 fwLeafWidget* fwRecursivelyFindLeafWidget(fwWidget* widget, int xpos, int ypos) {
@@ -47,13 +30,13 @@ fwLeafWidget* fwRecursivelyFindLeafWidget(fwWidget* widget, int xpos, int ypos) 
 	int row = frame->getRowCount() - 1, col = frame->getColumnCount() - 1;
 
 	for (int i = 1, rows = frame->getRowCount(); i < rows; ++i) {
-		if (frame->getYPos(i) > ypos) {
+		if (frame->getRowYPos(i) > ypos) {
 			row = i - 1;
 			break;
 		}
 	}
 	for (int i = 1, cols = frame->getColumnCount(); i < cols; ++i){
-		if (frame->getXPos(i) > xpos) {
+		if (frame->getColumnXPos(i) > xpos) {
 			col = i - 1;
 			break;
 		}