+++ fwWidget.cpp
@@ -3,10 +3,11 @@
 
 void fwWidget::attachTo(fwFrame* frame, int row, int column)
 {
-	parent = frame;
-	depth = frame->depth + 1;
-	pos[0] = frame->getXPos(column);
-	pos[1] = frame->getYPos(row);
-	size[0] = frame->getWidth(column);
-	size[1] = frame->getHeight(row);
+	m_parent = frame;
+	m_owner = frame->m_owner;
+	m_depth = frame->m_depth + 1;
+	m_pos[0] = frame->getColumnXPos(column);
+	m_pos[1] = frame->getRowYPos(row);
+	m_size[0] = frame->getColumnWidth(column);
+	m_size[1] = frame->getRowHeight(row);
 }